+++ CHANGELOG.md
@@ -11,7 +11,8 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 * ImageGallery
    * Rectangle images to be 16:9 ratio.
    * Fix 3 images to a row. Show grid layout for rest
-
+* ServiceMarketplace
+   * Add short description to card
 ## 1.0.0-RC10
 
 ### Change